requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-jsonlibrary
selenium
requests
openpyxl
